lab08/pyproject.toml
[project]
name = "lab08"
version = "1.0.0"
description = "Particle Swarm Optimization and Ant Colony Optimization"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pygad>=2.18.0",
    "numpy>=1.20.0",
    "matplotlib>=3.4.0",
    "networkx>=2.6.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.0",
    "pillow>=10.0.0",
    "pyswarms>=1.3.0",
    "aco>=0.2.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
    "pre-commit>=4.3.0",
    "commitizen>=4.13.9",
    "python-dotenv>=1.1.0",
    "pyright[nodejs]>=1.1.408",
    "ruff>=0.15.5",
]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [
	".venv",
	".pytest_cache",
	".ruff_cache",
	"__pycache__",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "PTH", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pyright]
include = ["task01/src", "task01/tests", "task02/src", "task02/tests", "task03/src", "task03/tests"]
exclude = [".venv", "__pycache__", "**/node_modules", "**/.*"]
typeCheckingMode = "standard"
venvPath = "./"
venv = ".venv"
reportMissingTypeStubs = "none"
reportUnknownVariableType = "none"
reportUnknownMemberType = "none"
reportUnknownArgumentType = "none"
reportUnknownParameterType = "none"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["task01/tests", "task02/tests", "task03/tests"]
addopts = ["--import-mode=importlib"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
major_version_zero = true
